+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 =========
+## Unreleased
+### Added
+- `TransformFilter::$matchCallback` callable to handle custom logic in determining of a transformer should be applied.
+
 ## 1.0.0-rc.1 - 2018-03-27
 ### Changed
 - Icons